Load Volbeat's palette right away upon entering the Biolum room
Since the Chinchou aren't visible right away when entering the room, and since
Volbeat's pallete was the reflection palette, Volbeat's palette wouldn't load
until the Chinchou were visible.
Since the Volbeat is visible from the door, flipping the palettes so that
Volbeat is the normal and Chinchou the reflection means that the two palettes
are loaded upon entering the room.